+++ client/src/Components/style/DinoItem.css
