+++ client/src/Components/style/DinoList.css
