+++ client/src/Components/style/component_images/card.png
